+++ opem/Functions.py
@@ -12,6 +12,18 @@ import webbrowser
 import sys
 
 
+def clear_screen():
+    """
+    Clear screen function.
+
+    :return: None
+    """
+    if sys.platform == "win32":
+        os.system('cls')
+    else:
+        os.system('clear')
+
+
 def integrate(y_vals, h):
     """
     Calculate integral with Simpson's Rule.